requirements.txt
flask
gunicorn
python-dotenv
GitPython
google-cloud-build
google-cloud-storage
